refactor(api): move endpoint to mission-browse controler
